+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add new component to enforce password strength
 * Add explicit error messages on 
 * Add push notifications
+* Add socket connection event on campus selection
 ### Fix
 * Fix infosec issue with X-Frame header
 * Fix push notifications